pyproject.toml
[project]
name = "emailagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "langsmith>=0.4.59",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
]
